port all existing bindings and tests
