+++ op-service/txintent/bindings/EventLogger.go
@@ -1,0 +1,9 @@
+package bindings
+
+import (
+	"github.com/ethereum-optimism/optimism/op-service/eth"
+)
+
+type EventLogger struct {
+	EmitLog func(topics []eth.Bytes32, data []byte) TypedCall[any] `sol:"emitLog"`
+}
